requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0

sentry-sdk[django]>=2.0,<3.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
